Handle password == null case, refactor some primitive types, set some relationships between entities, add create methods for entity managers, '

+++ src/main/java/com/areg/project/QuizConstants.java
@@ -5,10 +5,10 @@
 package com.areg.project;
 
 public class QuizConstants {
-    public static final int NumberOfRounds = 10;
-    public static final int RoundOptions = 4;
-    public static final int RoundTimeoutSeconds = 20;
-    public static final int PasswordSaltSize = 4;
+    public static final short NumberOfRounds = 10;
+    public static final byte RoundOptions = 4;
+    public static final byte RoundTimeoutSeconds = 20;
+    public static final byte PasswordSaltSize = 4;
     public static final String SecretKeyAlgorithm = "PBKDF2WithHmacSHA1";
     public static final String RNGAlgorithm = "SHA1PRNG";
 }
